refactor(dialog): extract CAboutDlg into standalone AboutDlg.{h,cpp}
Moves the embedded About dialog class out of ColorCopDlg.cpp into its own
compilation units (AboutDlg.h / AboutDlg.cpp). Removes duplicate class
definition and OnInitDialog implementation from ColorCopDlg.cpp. Updates
vcxproj and filters to include the new files.

No functional changes; behavior remains identical.
